minor refactors for restack
