+++ demos/react/src/App.tsx
@@ -1,5 +1,4 @@
 import { useHappyWallet } from "@happy.tech/react"
-import CallBatchDemo from "./demo-components/CallBatchDemo"
 import ChainSwitchingDemo from "./demo-components/ChainSwitchingDemo"
 import ConnectionDemo from "./demo-components/ConnectionDemo"
 import SessionKeyDemo from "./demo-components/SessionKeyDemo"
@@ -25,7 +24,6 @@ function App() {
                 <ChainSwitchingDemo />
                 <WalletFunctionalityDemo />
                 <SessionKeyDemo />
-                <CallBatchDemo />
             </div>
         </main>
     )
